+++ src/managers/CommunityManager.ts
@@ -1,4 +1,4 @@
-import axios from "axios"
+import fetch from "isomorphic-fetch"
 import { ManagerOptions, RequestConfig } from "../types/types"
 import { CommunityConfig, SetCommunityConfig, Community, ApiID } from "../types/apitypes"
 import BaseManager from "./BaseManager"
@@ -18,7 +18,7 @@ export default class CommunityManager extends BaseManager<Community> {
 			const cached = this.cache.get(communityId)
 			if (cached) return cached
 		}
-		const fetched = (await axios.get(`${this.apiurl}/communities/getid?id=${strictUriEncode(communityId)}`)).data
+		const fetched = await fetch(`${this.apiurl}/communities/getid?id=${strictUriEncode(communityId)}`).then(c=>c.json())
 
 		if (!fetched) return null // return null if the fetch is empty
 		if (fetched.error) throw new GenericAPIError(`${fetched.error}: ${fetched.description}`)
@@ -26,7 +26,7 @@ export default class CommunityManager extends BaseManager<Community> {
 		return fetched
 	}
 	async fetchAll(cache=true): Promise<Community[]> {
-		const allCommunities = (await axios.get(`${this.apiurl}/communities/getall`)).data
+		const allCommunities = await fetch(`${this.apiurl}/communities/getall`).then(c=>c.json())
 
 		if (allCommunities.error) throw new GenericAPIError(`${allCommunities.error}: ${allCommunities.description}`)
 
@@ -43,7 +43,7 @@ export default class CommunityManager extends BaseManager<Community> {
 		return null
 	}
 	async fetchConfig(guildId: string): Promise<CommunityConfig|null> {
-		const config = (await axios.get(`${this.apiurl}/communities/getconfig?guildId=${strictUriEncode(guildId)}`)).data
+		const config = await fetch(`${this.apiurl}/communities/getconfig?guildId=${strictUriEncode(guildId)}`).then(c=>c.json())
 		
 		if (config.error) throw new GenericAPIError(`${config.error}: ${config.description}`)
 		if (!config || !config.guildId) return null
@@ -51,9 +51,11 @@ export default class CommunityManager extends BaseManager<Community> {
 	}
 	async setConfig(config: SetCommunityConfig, reqConfig: RequestConfig = {}): Promise<CommunityConfig> {
 		if (!this.apikey && !reqConfig.apikey) throw new NoApikeyError()
-		const update = (await axios.post(`${this.apiurl}/communities/setconfig`, config, {
+		const update = await fetch(`${this.apiurl}/communities/setconfig`, {
+			method: "POST",
+			body: JSON.stringify(config),
 			headers: { "apikey": this.apikey || reqConfig.apikey, "content-type": "application/json" },
-		})).data
+		}).then(u=>u.json())
 		if (update.error) {
 			if (update.description === "API key is wrong") throw new AuthenticationError()
 			throw new GenericAPIError(`${update.error}: ${update.description}`)