+++ src/managers/InfoManager.ts
@@ -1,4 +1,4 @@
-import axios from "axios"
+import fetch from "isomorphic-fetch"
 import { ManagerOptions } from "../types/types"
 import { Webhook } from "../types/apitypes"
 import BaseManager from "./BaseManager"
@@ -12,24 +12,27 @@ export default class InfoManager extends BaseManager<Webhook> {
 		this.apiurl = apiurl
 	}
 	async addWebhook(webhookid: string, webhooktoken: string, guildid: string): Promise<Webhook> {
-		const add = (await axios.post(`${this.apiurl}/informatics/addwebhook`, {
-			id: webhookid,
-			token: webhooktoken,
-			guildId: guildid
-		}, {
-			headers: { "content-type": "application/json" },
-		})).data
-		return add
-	} 
-	async removeWebhook(webhookid: string, webhooktoken: string, guildid: string): Promise<Webhook> {
-		const add = (await axios.delete(`${this.apiurl}/informatics/removewebhook`, {
-			data: {
+		const add = await fetch(`${this.apiurl}/informatics/addwebhook`, {
+			method: "POST",
+			body: JSON.stringify({
 				id: webhookid,
 				token: webhooktoken,
 				guildId: guildid
-			},
+			}),
 			headers: { "content-type": "application/json" },
-		})).data
+		}).then(w=>w.json())
+		return add
+	} 
+	async removeWebhook(webhookid: string, webhooktoken: string, guildid: string): Promise<Webhook> {
+		const add = await fetch(`${this.apiurl}/informatics/removewebhook`, {
+			method: "DELETE",
+			body: JSON.stringify({
+				id: webhookid,
+				token: webhooktoken,
+				guildId: guildid
+			}),
+			headers: { "content-type": "application/json" },
+		}).then(w=>w.json())
 		return add
 	} 
 }