+++ src/managers/ProfileManager.ts
@@ -1,4 +1,4 @@
-import axios from "axios"
+import fetch from "isomorphic-fetch"
 import { ManagerOptions } from "../types/types"
 import { ApiID, Profile, Report } from "../types/apitypes"
 import BaseManager from "./BaseManager"
@@ -13,11 +13,13 @@ export default class ProfileManager extends BaseManager<Report> {
 		this.apiurl = apiurl
 	}
 	async fetchCommunity(playername: string, communityId: ApiID): Promise<Profile> {
-		const fetched = (await axios.get(`${this.apiurl}/profiles/getcommunity?playername=${strictUriEncode(playername)}&communityId=${strictUriEncode(communityId)}`)).data
+		const fetched = await fetch(`${this.apiurl}/profiles/getcommunity?playername=${strictUriEncode(playername)}&communityId=${strictUriEncode(communityId)}`)
+			.then(o=>o.json())
 		return fetched
 	}
 	async fetchAll(playername: string): Promise<Profile[]> {
-		const fetched = (await axios.get(`${this.apiurl}/profiles/getall?playername=${strictUriEncode(playername)}`)).data
+		const fetched = await fetch(`${this.apiurl}/profiles/getall?playername=${strictUriEncode(playername)}`)
+			.then(o=>o.json())
 		return fetched
 	}
 }