+++ src/managers/RevocationManager.ts
@@ -1,4 +1,4 @@
-import axios from "axios"
+import fetch from "isomorphic-fetch"
 import { ManagerOptions } from "../types/types"
 import { ApiID, Revocation } from "../types/apitypes"
 import BaseManager from "./BaseManager"
@@ -18,13 +18,15 @@ export default class RevocationManager extends BaseManager<Revocation> {
 		return null
 	}
 	async fetchRevocations(playername: string, communityId: string, cache = true): Promise<Revocation[]> {
-		const revocations = (await axios.get(`${this.apiurl}/revocations/getrevocations?playername=${strictUriEncode(playername)}&communityId=${strictUriEncode(communityId)}`)).data
+		const revocations = await fetch(`${this.apiurl}/revocations/getrevocations?playername=${strictUriEncode(playername)}&communityId=${strictUriEncode(communityId)}`)
+			.then(r=>r.json())
 		if (!revocations || !revocations[0]) return null
 		if (cache) revocations.forEach((revocation: Revocation) => this.add(revocation))
 		return revocations
 	}
 	async fetchAllRevocations(playername: string, cache = true): Promise<Revocation[]> {
-		const revocations = (await axios.get(`${this.apiurl}/revocations/getallrevocations?playername=${strictUriEncode(playername)}`)).data
+		const revocations = await fetch(`${this.apiurl}/revocations/getallrevocations?playername=${strictUriEncode(playername)}`)
+			.then(r=>r.json())
 		if (!revocations || !revocations[0]) return null
 		if (cache) revocations.forEach((revocation: Revocation) => this.add(revocation))
 		return revocations